Add CMake -DWITH_PIC option for static libraries
The static library is linked with a Perl XS shared object
and in that case we need to compile as position independent code.

In the general case however, this isn't necessary, thus the option
defaults to -DWITH_PIC=OFF. This doesn't affect shared libraries,
which are now always built as PIC.

Also adds a test that tests linking against the static library.